Change slab size related vars from uint64 to uint32
Currently, some slab size related variables, such as targetThreshold,
minThreshold, maxThreshold, are uint64.  However, ArraySlabHeader.size
and MapSlabHeader.size are uint32.  This causes frequent type casting
when comparing slab size with target slab size.

This commit refactors slab size related variables to use uint32 to
simplify code and be explicit about the supported max slab size
(math.MaxUint32).

This commit also modifies some exported function return types to
be consistent.

+++ array_verify.go
@@ -275,7 +275,7 @@ func (v *arrayVerifier) verifyDataSlab(
 		}
 
 		// Verify element size <= inline size
-		if e.ByteSize() > uint32(maxInlineArrayElementSize) {
+		if e.ByteSize() > maxInlineArrayElementSize {
 			return 0, nil, nil, NewFatalError(fmt.Errorf("data slab %s element %s size %d is too large, want < %d",
 				id, e, e.ByteSize(), maxInlineArrayElementSize))
 		}
@@ -284,7 +284,7 @@ func (v *arrayVerifier) verifyDataSlab(
 		case SlabIDStorable:
 			// Verify not-inlined element > inline size, or can't be inlined
 			if v.inlineEnabled {
-				err = verifyNotInlinedValueStatusAndSize(value, uint32(maxInlineArrayElementSize))
+				err = verifyNotInlinedValueStatusAndSize(value, maxInlineArrayElementSize)
 				if err != nil {
 					return 0, nil, nil, err
 				}
